♻️ refactor: common 컴포넌트 확장자 변경
