--- src/components/404/index.tsx
@@ -1,1 +1,0 @@
-export {};
