+++ src/components/common/ProfileImage/index.ts
@@ -1,0 +1,3 @@
+import ProfileImage from "./ProfileImage";
+
+export default ProfileImage;
