+++ src/components/common/SearchBar/style.ts
