+++ src/components/common/SkillIcon/style.ts
